PLATFORM/backend/bracelet-lib/pyproject.toml
[tool.poetry]
name = "bracelet-lib"
version = "0.1.0"
description = "Used by Digital bracelet Apis"
license = "Proprietary"
authors = ["Aprendiendo <depablodiegouma@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["email-validator"], version = "^1.9.0"}
email-validator = "^1.1.1"
sqlalchemy = "^1.4.36"
orjson = "^3.6.7"
asyncpg = "0.27.0"
psycopg2-binary = "^2.8.6"
sqlparse = "^0.4.1"
aioredis = "^1.3.1"
sentry-sdk = "^1.5.2"
aiosmtplib = "^1.1.6"
Mako = "^1.1.6"
PyJWT = "^2.4.0"
Numpy = "^1.9.1"
pyedflib = "^0.1.37"
databases = {git = "https://github.com/skuda/databases.git"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pydevd-pycharm = "^221.4165.171"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
